+++ application/app.py
@@ -1,15 +1,5 @@
 from dataclasses import dataclass
-from model.direction import Orientation
-from model.timber import Cutted2BY4
-from utility.draw import DrawIT
-from pathlib import Path
 import yaml
-import copy
-
-frame_photo_path = Path(__file__).parent / "frame.png" 
-config_path = Path(__file__).parent / "config.yml" 
-td = DrawIT(frame_photo_path.as_posix())
-
 
 def load_config(file_path: str):
     with open(file_path, 'r') as stream:
@@ -70,55 +60,3 @@ class App:
     def __get_wall_detailed_info(self, file_path):
         rt = load_config(file_path)
         return rt["parts"]
-
-
-
-class DrawAPP(App):
-    def __init__(self, file_path) -> None:
-        super().__init__(file_path)
-        self.current_move = 0
-
-    def draw(self, td: DrawIT, plates):
-        for instance in self.instances:
-            instance.draw(td)
-        # draw plates
-        for plate in plates:
-            td.prepare(plate)
-
-        td.draw_it()
-
-    def get_start_point(self):
-        return self.instances[0].get_area.a_cord.x
-
-    def get_end_point(self):
-        return self.instances[-1].get_area.b_cord.x
-
-    def execute(self):
-        super().execute()
-        # move first
-        for instance in self.instances:
-            instance.move_right(self.current_move)
-            self.current_move = self.current_move + instance.get_area.b_cord.x - instance.get_area.a_cord.x
-        
-        # after move, then get the location and get the end point
-        plate_size = self.get_end_point() - self.get_start_point()
-        plates = self.add_plates(plate_size, self.wall_global_info.floor_height)
-
-        #draw
-        self.draw(td, plates)
-
-    def add_plates(self, plate_size, floor_height):
-        bottom_plate = Cutted2BY4(plate_size, Orientation.HORIZONTAL)
-
-        top_plate = copy.copy(bottom_plate)
-        top_plate.move_up(Cutted2BY4.HEIGHT + floor_height)
-        
-        return [top_plate, bottom_plate]
-
-
-def main():
-    app = DrawAPP(config_path)
-    app.execute()
-    
-if __name__ == "__main__":
-    main()
